+++ extensions/cornerstone-dicom-sr/src/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import getSopClassHandlerModule from './getSopClassHandlerModule';
 import { srProtocol } from './getHangingProtocolModule';
+import onModeEnter from './onModeEnter';
 import getCommandsModule from './commandsModule';
 import preRegistration from './init';
 import { id } from './id.js';
@@ -8,8 +9,6 @@ import toolNames from './tools/toolNames';
 import hydrateStructuredReport from './utils/hydrateStructuredReport';
 import createReferencedImageDisplaySet from './utils/createReferencedImageDisplaySet';
 import Enums from './enums';
-import { ViewportActionButton } from '@ohif/ui';
-import i18n from '@ohif/i18n';
 
 const Component = React.lazy(() => {
   return import(/* webpackPrefetch: true */ './components/OHIFCornerstoneSRViewport');
@@ -32,29 +31,7 @@ const dicomSRExtension = {
    */
   id,
 
-  onModeEnter({ servicesManager }) {
-    const { toolbarService } = servicesManager.services;
-
-    toolbarService.addButtons([
-      {
-        // A base/default button for loading measurements. It is added to the toolbar below.
-        // Customizations to this button can be made in the mode or by another extension.
-        // For example, the button label can be changed and/or the command to clear
-        // the measurements can be dropped.
-        id: 'loadSRMeasurements',
-        component: props => (
-          <ViewportActionButton {...props}>{i18n.t('Common:LOAD')}</ViewportActionButton>
-        ),
-        props: {
-          commands: ['clearMeasurements', 'loadSRMeasurements'],
-        },
-      },
-    ]);
-
-    // The toolbar used in the viewport's status bar. Modes and extensions can further customize
-    // it to optionally add other buttons.
-    toolbarService.createButtonSection('loadSRMeasurements', ['loadSRMeasurements']);
-  },
+  onModeEnter,
 
   preRegistration,
